requirements.txt
requests
selenium
webdriver-manager
ruff
dspy-ai
google-generativeai
python-dotenv
Pillow
